+++ client/src/components/page-template/PageTemplate.js
@@ -27,6 +27,7 @@ import {
   ServerProxy,
   DnsServices,
   IbmCloudSysdigSecure,
+  IbmDb2,
 } from "@carbon/icons-react";
 import f5 from "../../images/f5.png";
 import { arraySplatIndex, contains, getObjectFromArray } from "lazy-z";
@@ -70,6 +71,7 @@ const navIcons = {
   ServerProxy: ServerProxy,
   DnsServices: DnsServices,
   IbmCloudSysdigSecure: IbmCloudSysdigSecure,
+  IbmDb2: IbmDb2,
   CBRIcon: CBRIcon,
   F5Icon: F5Icon,
 };